Seed training run in train_basis.py; move set_seed into utils
